requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
idna==3.3
PyJWT==2.4.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
